feat(blast-radius): promote Impact view into @repowise-dev/ui
Move the blast-radius analyzer (searchable file picker, hotspot chips,
depth control, risk gauge + impact graph + detail tables) into
packages/ui behind an injected ImpactAdapter. The reviewer panel is a
host-supplied slot that no-ops when absent (graceful degradation). Web
becomes a thin host wiring its /api client + reviewer fetch; the
rendered view is unchanged. Adds view-level behavioral tests.
